ai/requirements.txt
biopython
requests
pandas
numpy
git+https://github.com/huggingface/transformers.git
lxml